+++ braintree/android_pay_card.py
@@ -31,6 +31,10 @@ class AndroidPayCard(Resource):
         return self.bin
 
     @property
+    def token(self):
+        return self.token
+
+    @property
     def commercial(self):
         return self.commercial